examples/agent/6-appsource/capital_agent/agent.py
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import os
from google.adk.agents import LlmAgent

MODEL_ID = os.getenv("MODEL_ID", "gemini-2.0-flash")


# Define a tool function
def get_capital_city(country: str) -> str:
    """Retrieves the capital city for a given country."""
    # Replace with actual logic (e.g., API call, database lookup)
    capitals = {"france": "Paris", "japan": "Tokyo", "canada": "Ottawa"}
    return capitals.get(country.lower(),
                        f"Sorry, I don't know the capital of {country}.")


capital_agent = LlmAgent(
    model=MODEL_ID,
    name="capital_agent",
    description="""Answers user questions about the capital
    city of a given country.""",
    instruction="""You are an agent that provides the capital
    city of a country.
    When a user asks for the capital of a country:
    1. Identify the country name from the user's query.
    2. Use the `get_capital_city` tool to find the capital.
    3. Respond clearly to the user, stating the capital city.
    Example Query: "What's the capital of France?"
    Example Response: "The capital of France is Paris."
    """,
    tools=[get_capital_city]  # Provide the function directly
)

root_agent = capital_agent
